Fix Vercel deployment error by moving utils.js out of API directory
Move utils.js from pages/api/ to lib/ to prevent Next.js from treating
it as an API endpoint. Update all imports accordingly.
