+++ lib/utils.js
@@ -144,3 +144,20 @@ export function formatDateET(isoDateString) {
     day: 'numeric' 
   });
 }
+
+/**
+ * Create an AbortSignal with timeout that works across Node.js versions
+ * @param {number} ms - Timeout in milliseconds
+ * @returns {AbortSignal} Abort signal that triggers after timeout
+ */
+export function createTimeoutSignal(ms) {
+  // Use AbortSignal.timeout if available (Node 16.14+)
+  if (typeof AbortSignal !== 'undefined' && typeof AbortSignal.timeout === 'function') {
+    return AbortSignal.timeout(ms);
+  }
+  
+  // Fallback for older Node.js versions
+  const controller = new AbortController();
+  setTimeout(() => controller.abort(), ms);
+  return controller.signal;
+}
