+++ package.json
@@ -2,6 +2,9 @@
   "name": "material-tailwind-project",
   "private": true,
   "version": "0.0.0",
+  "engines": {
+    "node": ">=18.17.0"
+  },
   "scripts": {
     "dev": "next dev",
     "build": "next build",
